add new script setup_image.py for base and final image generation
This script runs in a Docker build step via RUN.

Along with it we also have a credentials file in yaml format, which
contains the container root and various db user passwords, so they
may be set in the final image build.

User creds have also been added to the default config.

The db primary image has been cleaned up a bit. A fixed server_id
and gtid_domain are now set, and db files are imported for the specified
wikdbs. Additionally there is a tiny amount of testing now of db-related
methods in the build.

We also now show build logs if --verbose is specified.