DO NOT USE
This adds the php-fpm setup, the mounted volume for the MediaWiki repo
checkout, and the ability to destroy containers and images by name just
as we can build them by name. It also adds some a MOTD for containers
and some html/php pages providing basic info but also serving
as tests that a given config works. And it fixes up the apt repo setups
for the various images. However it is guaranteed to be broken. Wait for the
next commit in a day or so, which should sort everything back out.